requirements.txt
certifi==2020.12.5
psycopg2-binary==2.8.6
python-dotenv==0.15.0
